requirements.txt
torch>=1.8.0
torchvision
timm
wandb
numpy
Pillow
requests
